Move FavoritesDialog inside Dialog directory
